Extract FutureWallet to separate file
