+++ universal-login-sdk/lib/index.ts
@@ -2,7 +2,8 @@ import UniversalLoginSDK from './api/sdk';
 export default UniversalLoginSDK;
 export {SdkConfig} from './config/SdkConfig';
 export {SdkSigner} from './api/SdkSigner';
-export {FutureWallet, BalanceDetails} from './api/FutureWalletFactory';
+export {BalanceDetails} from './api/FutureWalletFactory';
+export {FutureWallet} from './api/wallet/FutureWallet';
 export {DeployingWallet} from './api/wallet/DeployingWallet';
 export {WalletService} from './core/services/WalletService';
 export {WalletStorage, SerializedWalletState, WalletState, SerializedDeployingWallet} from './core/models/WalletService';
